+++ src/main/java/com/codetaylor/mc/dropt/modules/dropt/rule/data/RuleDropItemMatchQuantity.java
@@ -1,0 +1,13 @@
+package com.codetaylor.mc.dropt.modules.dropt.rule.data;
+
+import net.minecraft.item.crafting.Ingredient;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class RuleDropItemMatchQuantity {
+
+  public transient List<Ingredient> _drops = new ArrayList<>();
+
+  public String[] drops = new String[0];
+}